Fix guns appearing to cycle bolt when wielded (#28756)
Adds a check during wielding to see if the code is running clientside, and if so skip the part responsible for creating the virtual items.

This is necessary because TrySpawnVirtualItem is blocked from running clientside, so trying to spawn the virtual items for wielding causes the client to always believe the wield has failed. This erroneous failure leads to the display of incorrect feedback until the server's successful wield attempt makes it to the client. The added check prevents wielding from failing in this way and therefore allows the client to behave as expected.

(cherry picked from commit 249e4a9c15c8e3960dd1cd77db4ad4090bf6609d)