Shield Reworks (#80)
## About the PR
Reopens (https://github.com/echo-station/echo-station/pull/37)

## Why / Balance
Shields are very brittle and will shatter way before the user goes into
a critical state, regardless if the user was using the supposedly
correct shield for the correct damage type. This PR ultimately seeks to
make shields a more useful tool with an ultimately longer shelf life,
with respect to their niche specialization.

Shields now are able to be stored in backpacks, consuming a 5x3 grid
(taller than wide). This idea is ultimately aimed towards user
ergonomics and quality of life, because without this change, you would
need to sacrifice a hand slot to bring a shield with you to any
environment.
Further justification: The Energy shield and the telescopic shield are
able to be stored in a backpack while deployed, and they consume a 4x4
grid, which is also the same size as a blood-red hardsuit and most large
weapons.
This change is also aimed towards a future weapon size rebalance which
will aim to convert most rifle sized weapons to 5x2 grid slots.

**Changelog**
:cl: ProphetCrow
- tweak: Shield stats have been modified