Added CentCom pAI (#96)
## About the PR
Added a CentCom pAI as requested in
[Issue75](https://github.com/echo-station/echo-station/issues/75).

## Why / Balance
Because it's funny and was requested.

As for balance, a CentCom pAI behaves almost the same as a normal pAI
except for its ability to electrocute everyone near them if they so
desire (once every 15 seconds). They can also give themselves All Access
for 30 seconds, making it so that whoever is holding them can go
wherever they want in the station! The also have access to almost all
radio channels.

## Technical details
Looks different, can electrocute, can turn on All Access, has access to
most radio channels, has a different name prefix when name changed.

## Media
![tmp](https://github.com/user-attachments/assets/e516fe96-a4f1-417e-9e06-e1bc35569ed1)

![tmp2](https://github.com/user-attachments/assets/4687ea52-1029-43b9-a039-486776462164)

![tmp3](https://github.com/user-attachments/assets/6fdcda59-461f-48e0-bf2e-2abe8b14cc79)

## Breaking changes

None that I'm aware of.

**Changelog**

:cl:
- add: Added a CentCom pAI.